Calculate fishing mortality under the PA leaf

These functions calculate fishing mortality based on biomass using different harvest control rule (HCR) approaches:

F_linear(B, Blim, Btrigger, Ftarget, Fmin = 0)

F_sigmoid(B, Blim, Btrigger, Ftarget, x50, k = 1, Fmin = 0)

F_lower(B, Blim, Btrigger, Ftarget, k = 1, Fmin = 0)

F_upper(B, Blim, Btrigger, Ftarget, k = 1, Fmin = 0)

Arguments

B

Biomass vector

Blim

Biomass limit reference point

Btrigger

Biomass trigger reference point

Ftarget

Target fishing mortality

Fmin

Minimum fishing mortality (default: 0)

x50

(for F_sigmoid) Proportion of the biomass range at which F = 0.5 * Ftarget

k

(for F_sigmoid) Shape parameter for the sigmoid function (default: 1)

Value

Numeric vector of fishing mortality values

Details

F_linear() implements a linear HCR

F_sigmoid() implements a sigmoid HCR with configurable inflection point (x50)

F_lower() sets x50 = 0.75 for the lower bound of the leaf as per Annex 4 in Com-SC RBMS WP 23-03 Rev 2

F_upper() sets x50 = 0.25 for the upper bound of the leaf as per Annex 4 in Com-SC RBMS WP 23-03 Rev 2