Toy Stock Data

A simulated dataset for demonstrating plotting functions in the NAFOdown R package. The dataset contains simulated fish stock dynamics over a 40-year period.

toy_stock

Format

A data frame with 40 rows and 13 variables:

year

Year of the simulation, ranging from 1981 to 2020.

biomass

Simulated biomass of the fish stock.

rec_index1

Simulated recruitment index 1.

rec_index2

Simulated recruitment index 2.

rec_index3

Simulated recruitment index 3.

TAC

Total allowable catch (TAC) for the fish stock.

catch

Simulated catch of the fish stock.

Fmsy

Fishing mortality rate at maximum sustainable yield.

Bmsy

Biomass at maximum sustainable yield.

relB

Relative biomass compared to 30% of Bmsy.

relB_lwr

Lower bound of relative biomass confidence interval.

relB_upr

Upper bound of relative biomass confidence interval.

relF

Relative fishing mortality rate compared to Fmsy.

relF_lwr

Lower bound of relative fishing mortality rate confidence interval.

relF_upr

Upper bound of relative fishing mortality rate confidence interval.

Details

The data were generated using a simple surplus production model.