Function for producing summary tables for NAFO documents

Produces summary tables for NAFO stock summary sheets

nafo_summary_table(comments = rep(NA, 5), status = rep(NA, 5), parse = FALSE)

Arguments

comments

A vector of length 5 elaborating on the status the status of each 'Convention General Principle'

status

A vector of length 5 describing the status of each 'Convention General Principle'. Must be one of: 'OK', 'Intermediate','Not accomplished', 'Unknown'

parse

Parse comments with markdown like syntax (experimental!)

Examples

nafo_summary_table(status = c("Not accomplished", "OK", "OK", "Intermediate", "Unknown"),
                   comments = c("Probability of *B~2019~* < *B~MSY~* = 97%",
                                "*F* < *F~MSY~*",
                                "Reference points defined",
                                "VME closures in effect, no specific measures",
                                "Cannot be evaluated"),
                   parse = TRUE)

Convention General Principles

Status

Comment/consideration

 

  

Restore or maintain at BMSY

●

Probability of B2019 < BMSY = 97%

●

OK

Eliminate overfishing

●

 F < FMSY

●

Intermediate

Apply Precautionary Approach

●

Reference points defined

●

Not accomplished

Minimise harmful impacts on living marine resources and ecosystems

●

VME closures in effect, no specific measures

●

Unknown

Preserve marine biodiversity

●

Cannot be evaluated